+++ src/commands/ITheme.ts
@@ -1,0 +1,14 @@
+/**
+ * Object that represents an installed theme
+ */
+export interface ITheme {
+    /** Id of the extension the theme belongs to */
+    extensionId: string;
+    /** Label of the theme.  Used as a friendly name of the theme */
+    label: string;
+    /** 
+     * Most theme extensions don't provide an id for each theme, but when they do, that id
+     * should be used to set the active theme.
+     */
+    themeId: string | null;
+}